chore: apply @mini-max changes for PR #10444
