+++ src/transpiler/src/emitter/index.ts
@@ -2,8 +2,7 @@ export type * from "./ast.js";
 export { emitBuiltinFunction, isBuiltinFunction } from "./builtins.js";
 export { wrapConditional, wrapConditionalBody, wrapRawBody } from "./code-wrapping.js";
 export { tryFoldConstantExpression } from "./constant-folding.js";
-export { GmlToJsEmitter } from "./emitter.js";
-export { emitJavaScript } from "./emitter-entrypoints.js";
+export { emitJavaScript, GmlToJsEmitter } from "./emitter.js";
 export { lowerEnumDeclaration } from "./enum-lowering.js";
 export { escapeTemplateText, isIdentifierLike, normalizeStructKeyText, stringifyStructKey } from "./js-string-utils.js";
 export { normalizeGmlNumericLiteral } from "./literal-normalization.js";
